Remove submit logic from click event handler (#79)
There is no need to have submit logic take place in the on click handler. It only deals with the mouse click event, *not* the submit event that will follow. 

Before this change, blocking the default action on the click event prevents a following submit event from taking place: `form.onsubmit` and other registered listeners waiting for a 'submit' will never be called.